backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
black==25.1.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distlib==0.3.9
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.0
elastic-transport==8.17.1
elasticsearch==8.17.2
email_validator==2.2.0
fastapi==0.115.11
filelock==3.17.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.9
idna==3.10
iniconfig==2.1.0
isort==6.0.1
msgpack==1.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pillow==11.1.0
pipenv==2024.4.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
pytest==8.3.5
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
ruff==0.11.2
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tqdm==4.67.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uv==0.5.31
uvicorn==0.34.0
virtualenv==20.29.2

